Use the window surface position in parent's coordinate for InsetsSource
Window frames is based on the display coordinate, but the surface
position should be based on the surface parent's coordinate.

On single display device, the parent is at (0,0) for most case, so it
doesn't cause any issue. However, on multi-display device, the system
and IME window can be placed in a DisplayArea that is not at (0,0).

Bug: 156355859
Test: manual: open IME on single and dual display device.
Change-Id: I42de9b8a12db56b2a9e0c2de0c1e1d7c5d0f25ba