Remove unnecessary catch statement
The IllegalFormatConversionException catch statement was added when
there was a change in a string that required a new translation in all
languages (in P). All translations should be properly updated now, so we
shouldn't need this catch statement. [If we do see this exception, we
should check the locale and update the string's translation.]

Test: manual
Fixes: 174014345
Change-Id: I790870cb90ca8f6904ef73f481928a7cdd2faa05