Introduce AudioProfile to describe the capabilities of audio ports.
For an audio port, the capabilities are currently described by flat
arrays of formats, channel masks and sample rates. This doesn't expose
the relationship between supported formats and channel masks/sample
rates.
In that case, AudioProfile is introduced which contains the audio
format and its associated sample rates and channel masks.

Test: atest AudioManagerTest
Bug: 25994457
Bug: 160352965
Change-Id: Ic99713e32a1585875001ee096eb27d6c9c5cb64f