Add VERSION_CODES constant for S
This is a subset of the commit ba60b3922df26a760272a78fa46a2527a62e68cd
from internal master.

Note: The CL used for the Merged-in tag is irrelevant to this change but
achieves the desired automerging effects.

Test: No
Merged-In: If5c3b92063b8d9fb3f9b6272eb2cfef448065380
Change-Id: I205766a281eb3956c3d122f5358fe5c984803e6e