Add ability to ignore gestures in FalsingManager
This gives other parts of the system the ability to temporarily
override the falsing manager, telling it to ignore a given gesture.

The intent is to give other parts of the system a chance to include
their own falsing information (such as whether a button was tapped or
not), giving the FalsingManager more context about what is going on.

As of right now, their is no way to feedback this context-aware
falsing information back to the FalsingManager. An API for this is
still being explored.

Bug: 172655679
Test: atest SystemUITests && manual
Change-Id: I87ee94a2386622ebff7580346325ae07e3bb5f5d