Prevent duplicate call to update movement bounds on rotation
When rotating the display while in PIP mode, there's 1 call
to reset the shelf offset back to zero which triggers a call
to PipController.updateMovementBounds(fromShelfAdj=true).
Then, another call is made immediately to
PipController.updateMovementBounds(fromRotation=true).

While both are useful when called independently, we now
avoid calling the first on rotation, otherwise it'll try to
offset the pre-rotation PIP bounds and use those bounds as
the post-rotation bounds instead of calculating the new bounds
using the rotated display and the previous snap fraction.

Bug: 174669093
Test: PIP position is preserved when rotating the home screen
Change-Id: I071ace219bd26a5c8078b60c77a8b2442470305c